Conflicts:
	portfolio_matching/target/m2e-wtp/web-resources/META-INF/maven/com.spring/ex/pom.properties

+++ portfolio_matching/src/main/java/com/spring/ex/vo/PortfolioVO.java
@@ -10,11 +10,10 @@ public class PortfolioVO {
 	private String portfolio_sCat;
 	private String portfolio_thumbUrl;
 	private String portfolio_Img;
-
 	private String portfolio_content;
 	private int portfolio_price;
 	private String portfolio_term;
-	private int portfolio_rating;
+	private double portfolio_rating;
 	private Date portfolio_regDate;
 
 	public int getPortfolio_id() {
@@ -97,11 +96,11 @@ public class PortfolioVO {
 		this.portfolio_term = portfolio_term;
 	}
 
-	public int getPortfolio_rating() {
+	public double getPortfolio_rating() {
 		return portfolio_rating;
 	}
 
-	public void setPortfolio_rating(int portfolio_rating) {
+	public void setPortfolio_rating(double portfolio_rating) {
 		this.portfolio_rating = portfolio_rating;
 	}
